Finished the backend code for signing in and up

+++ src/main/java/com/gabrielpf/wodfeeder/model/auth/User.java
@@ -1,7 +1,6 @@
 package com.gabrielpf.wodfeeder.model.auth;
 
 import com.gabrielpf.wodfeeder.vo.SignUpUserVO;
-import com.gabrielpf.wodfeeder.vo.UserInVO;
 import lombok.Data;
 
 import javax.persistence.Column;
@@ -18,7 +17,7 @@ public class User {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id")
-	private int id;
+	private Long id;
 
 	@Column(name = "username", nullable = false, unique = true)
 	private String username;
